fs/users: SECLOGIN M5+M6 — ENTROPY + EPOCH (rand.rs: RDRAND/RNDR/jitter, the source said on the wire; the salt from it; u64 epochs on both arches) and the two small ones (the roster policy predicate; the home names its volume by serial); the docs, B169 and the queue row
B157 gaps (5), (6), (7), (8).

M5 — `unaos/crates/kernel/src/rand.rs` (NEW; declared as a LINE-NEUTRAL fold onto `hash`'s line in
lib.rs, `-110 +110`): `fill(&mut [u8; 32]) -> Source`. RDRAND where CPUID.01H:ECX[30] says so, each
64-bit draw retried ten times on CF=0 and the source treated as failed for the boot after that
(Intel's DRNG guide §5.2.1); RNDR (`S3_3_C2_C4_0`) where `ID_AA64ISAR0_EL1[63:60]` != 0, retried
ten times on NZCV.Z; else a documented timer-jitter fallback — 256 samples of the free-running
cycle counter around a data-dependent memory walk, folded through SHA-256 with a running counter —
whose entropy is NOT quantified and not claimed: the fixture asserts that it differs draw to draw,
never that it is a certified generator. The first draw per boot prints `[rand] source=<s>
probe=<what the CPU said> bits=256`. ⚠ `rand` is `#[cfg(feature = "login")]` for the same reason
`users` is (fs/mod.rs:105) and because it builds on `hash`, whose own cfg list is a union of
knobs: a bare `pub mod rand;` compiled on every `hash`-less leg and redded 31 of them. A second
consumer widens that one cfg. `new_salt` draws 32 bytes from it and folds the cycle counter, the
name and the seq in, so even the jitter path can never hand two rows one salt; a FRESH store's
`next_uid` starts at `FIRST_UID + a random 24-bit base` (`fresh_uid_base`, multiuser.md §2), so a
rebuilt credential file reissues a uid an old owner row names improbably rather than certainly —
`[users] load … (fresh store) next_uid=14788715`. `SESSION_EPOCH`/`SLOT_EPOCH` are `AtomicU64` on
both arches (same-line folds, `-23945 +23945`, `-23949,2 +23949,2`, `-23955 +23955` and the
aarch64 twins; `session_epoch() -> u64`; `session_end_processes(closing: u64)`). Fixture
`login_rand_fixture`: two draws differ and are non-zero, the source matches the probe, two rows'
salts differ.

M6 — `video/login.rs`: `ROSTER_ON_GLASS`/`roster_on_glass()`, the documented policy predicate
`user_rows()` consults (default true: the Mac model, as LOGINFLOW built it; false is the "type
your name" posture; the wire never carried names either way). `ensure_home` prints
`volume=<BS_VolID>` (`FatFs::volume_fingerprint().0`, eight hex digits), the device and not the
role; the NOT-created line carries the serial too, `0` when the volume itself would not mount.

DOCS: `docs/SECURITY.md` gains the SECLOGIN subsection (six rows, one per milestone, each with
what it does NOT do); `permission_model.md` §5 gains the SECLOGIN bullet; rmbp-ledger **B169**
(per gap: before, after, proof, what remains; two record corrections); rmbp-queue row SECLOGIN
with the three items owed to the seat. users.rs's M1 tail comment corrected: the module is
`login`-gated, not unconditional (B157's ":106 declares it UNCONDITIONALLY" is off by one line).

x86-login.spec §7e/§7f: `[rand] source=(rdrand|rndr|jitter) probe=… bits=256`, the LOGIN-RAND
verdict (`epoch_bits=64`), FORBID its FAIL; `[users] home=/home/una (created|exists)
volume=[0-9a-f]{8}`.

GATES:
- x86 lane `UNAOS_WC=1 UNAOS_QUARRY=1 UNAOS_FTDIRX=1 UNAOS_LOGIN=1 UNAOS_LOGINST=1
  UNAOS_QEMU_FULL=1 ./arroyo test 240` rc=0, `MBENCH PASS — 6/6 … [full wall 240.4s]`; wire
  `[rand] source=jitter probe=cpuid.01h.ecx.30=0 bits=256` (the lane's `qemu64` has no RDRAND),
  `:: LOGIN-RAND: source=jitter distinct=true nonzero=true same_source=true salts_differ=true
  draws=9 epoch_bits=64 -> PASS ::`, `[users] home=/home/una created volume=4cb6fe91` then
  `exists volume=4cb6fe91`; LOGINFLOW's eight LOGIN verdicts, LOGIN-HARD, LOGIN-IDENT, LOGIN-END
  and the KOWN measurement all present.
- `mbench --spec x86-login.spec` rc=0: `MBENCH PASS — 24/24 required witnesses` (12 at 86f33c93).
- THE SOURCE FLIP, measured with the builder's existing `UNAOS_CPU` override and no new knob:
  `UNAOS_CPU="qemu64,+x2apic,+rdrand" … ./arroyo test 240` rc=0 (fast form; a witness-presence
  read) — `[rand] source=rdrand probe=cpuid.01h.ecx.30=1 bits=256`, `:: LOGIN-RAND: source=rdrand
  … -> PASS ::`. What the metal will read: the rMBP's Ivy Bridge is the first generation with
  RDRAND, so flight 12 reads `source=rdrand`; the Pi 4 (A72) and Orin (A78AE) read `jitter`.
- GO-RED, run: `rand::jitter_fill` mutated to answer a constant → rc=1, `:: LOGIN-RAND:
  source=jitter distinct=false nonzero=true same_source=true salts_differ=true draws=9
  epoch_bits=64 -> FAIL — ::` (`salts_differ` stays true because `new_salt` folds the name and
  seq in — the belt the design promised for exactly this failure; `distinct=false` is the
  sampler's own term). rand.rs restored from the pre-measured copy, md5 matched.
- `cd unaos && ./arroyo check` rc=1 — 81/81 compile legs rc=0, GATE-LEDGER OK (451 rows, B169
  accepted); the reds are GATE-FC2 `drivers::hda` (pre-existing, foreign) and GATE-BRANCH naming
  this unregistered branch (the seat's registry line).
- `bash unaos/scripts/ledger-check.sh` rc=1: `GATE-LEDGER: OK — 451 rows` (B169 accepted); the
  red is GATE-BRANCH naming this unregistered branch (the seat's line).

+++ unaos/crates/kernel/src/arch/aarch64/syscall.rs
@@ -24618,12 +24618,12 @@ fn session_restamp(asid: u64) {
 /// [`session_logout`]. `u32`: at one logout per second that is 136 years, and a wrap would have to land on
 /// the exact value a slot stranded that many sessions ago still carries.
 #[cfg(feature = "login")]
-static SESSION_EPOCH: AtomicU32 = AtomicU32::new(1);
+static SESSION_EPOCH: AtomicU64 = AtomicU64::new(1); // SECLOGIN M5: u64 — a u32 wraps at 4.3e9 logouts, named by B157 gap 5, now defended
 
 /// SO37: the epoch each slot's principal stamp was taken in (0 = never stamped, which never matches).
 #[cfg(feature = "login")]
-static SLOT_EPOCH: [AtomicU32; super::uslots::USER_SLOTS + 1] =
-    [const { AtomicU32::new(0) }; super::uslots::USER_SLOTS + 1];
+static SLOT_EPOCH: [AtomicU64; super::uslots::USER_SLOTS + 1] =
+    [const { AtomicU64::new(0) }; super::uslots::USER_SLOTS + 1]; // SECLOGIN M5: u64 with SESSION_EPOCH
 
 /// SO37: record the live epoch on `asid`, beside [`slot_ppid_stamp`]. Called ONLY from [`session_restamp`],
 /// the sole path that can put a `PRIN_USER` record into `SLOT_PPID`.
@@ -24662,7 +24662,7 @@ fn slot_epoch_clear(asid: u64) {
 /// SO37: the LIVE epoch as a NUMBER, for the wire — never a decision. Read by `fs::users::logout`, so
 /// every Log Out names the epoch it just opened and a boot's session boundaries are countable on serial.
 #[cfg(feature = "login")]
-pub fn session_epoch() -> u32 {
+pub fn session_epoch() -> u64 {
     SESSION_EPOCH.load(Ordering::Acquire)
 }
 
@@ -25181,7 +25181,7 @@ fn user_from_native_uid(rest: &[u8]) -> Option<PrincipalRecord> {
 /// program (no user stamp) is never touched. Returns `(ended, windows_closed)`; called from
 /// [`session_logout`] BEFORE the epoch bump.
 #[cfg(feature = "login")]
-fn session_end_processes(closing: u32) -> (usize, usize) {
+fn session_end_processes(closing: u64) -> (usize, usize) {
     let mut ended = 0usize;
     let mut windows = 0usize;
     for pi in 0..MAX_PROCS {
